+++ pdfminer/pdftypes.py
@@ -27,7 +27,7 @@ LITERALS_ASCIIHEX_DECODE = (LIT('ASCIIHexDecode'), LIT('AHx'))
 LITERALS_RUNLENGTH_DECODE = (LIT('RunLengthDecode'), LIT('RL'))
 LITERALS_CCITTFAX_DECODE = (LIT('CCITTFaxDecode'), LIT('CCF'))
 LITERALS_DCT_DECODE = (LIT('DCTDecode'), LIT('DCT'))
-
+LITERALS_JBIG2_DECODE = (LIT('JBIG2Decode'),)
 
 ##  PDF Objects
 ##
@@ -275,6 +275,8 @@ class PDFStream(PDFObject):
                 # This is probably a JPG stream - it does not need to be decoded twice.
                 # Just return the stream to the user.
                 pass
+            elif f in LITERALS_JBIG2_DECODE:
+                pass                
             elif f == LITERAL_CRYPT:
                 # not yet..
                 raise PDFNotImplementedError('/Crypt filter is unsupported')