+++ tests/test_tools_pdf2txt.py
@@ -2,72 +2,71 @@ import os
 from shutil import rmtree
 from tempfile import NamedTemporaryFile, mkdtemp
 
-import nose
-
 import tools.pdf2txt as pdf2txt
+from helpers import absolute_sample_path
 
 
-def full_path(relative_path_to_this_file):
-    this_file_dir = os.path.dirname(os.path.abspath(__file__))
-    abspath = os.path.abspath(os.path.join(this_file_dir, relative_path_to_this_file))
-    return abspath
-
-
-def run(datapath, filename, options=None):
-    i = full_path(datapath + filename + '.pdf')
-    o = full_path(filename + '.txt')
-    if options:
-        s = 'pdf2txt -o%s %s %s' % (o, options, i)
-    else:
-        s = 'pdf2txt -o%s %s' % (o, i)
-    pdf2txt.main(s.split(' ')[1:])
+def run(sample_path, options=None):
+    absolute_path = absolute_sample_path(sample_path)
+    with NamedTemporaryFile() as output_file:
+        if options:
+            s = 'pdf2txt -o %s %s %s' % (output_file.name, options, absolute_path)
+        else:
+            s = 'pdf2txt -o %s %s' % (output_file.name, absolute_path)
+        pdf2txt.main(s.split(' ')[1:])
 
 
 class TestDumpPDF():
 
-    def test_1(self):
-        run('../samples/', 'jo')
-        run('../samples/', 'simple1')
-        run('../samples/', 'simple2')
-        run('../samples/', 'simple3')
-        run('../samples/','sampleOneByteIdentityEncode')
+    def test_jo(self):
+        run('jo.pdf')
 
-    def test_2(self):
-        run('../samples/nonfree/', 'dmca')
+    def test_simple1(self):
+        run('simple1.pdf')
 
-    def test_3(self):
-        run('../samples/nonfree/', 'f1040nr')
+    def test_simple2(self):
+        run('simple2.pdf')
 
-    def test_4(self):
-        run('../samples/nonfree/', 'i1040nr')
+    def test_simple3(self):
+        run('simple3.pdf')
 
-    def test_5(self):
-        run('../samples/nonfree/', 'kampo')
+    def test_sample_one_byte_identity_encode(self):
+        run('sampleOneByteIdentityEncode.pdf')
 
-    def test_6(self):
-        run('../samples/nonfree/', 'naacl06-shinyama')
+    def test_nonfree_175(self):
+        """Regression test for https://github.com/pdfminer/pdfminer.six/issues/65"""
+        run('nonfree/175.pdf')
 
-    # this test works on Windows but on Linux & Travis-CI it says
-    # PDFSyntaxError: No /Root object! - Is this really a PDF?
-    # TODO: Find why
-    """
-    def test_7(self):
-        run('../samples/contrib/','stamp-no')
-    """
+    def test_nonfree_dmca(self):
+        run('nonfree/dmca.pdf')
 
-    def test_8(self):
-        run('../samples/contrib/', '2b', '-A -t xml')
+    def test_nonfree_f1040nr(self):
+        run('nonfree/f1040nr.pdf')
 
-    def test_9(self):
-        run('../samples/nonfree/', '175')  # https://github.com/pdfminer/pdfminer.six/issues/65
+    def test_nonfree_i1040nr(self):
+        run('nonfree/i1040nr.pdf')
 
-    def test_10(self):
-        run('../samples/scancode/', 'patchelf')  # https://github.com/euske/pdfminer/issues/96
+    def test_nonfree_kampo(self):
+        run('nonfree/kampo.pdf')
+
+    def test_nonfree_naacl06_shinyama(self):
+        run('nonfree/naacl06-shinyama.pdf')
+
+    def test_nlp2004slides(self):
+        run('nonfree/nlp2004slides.pdf')
+
+    def test_contrib_2b(self):
+        run('contrib/2b.pdf', '-A -t xml')
+
+    def test_scancode_patchelf(self):
+        """Regression test for # https://github.com/euske/pdfminer/issues/96"""
+        run('scancode/patchelf.pdf')
 
 
 class TestDumpImages(object):
 
-    def extract_images(self, input_file):
+    @staticmethod
+    def extract_images(input_file):
         output_dir = mkdtemp()
         with NamedTemporaryFile() as output_file:
             commands = ['-o', output_file.name, '--output-dir', output_dir, input_file]
@@ -81,13 +80,25 @@ class TestDumpImages(object):
 
         Regression test for: https://github.com/pdfminer/pdfminer.six/issues/131
         """
-        image_files = self.extract_images(full_path('../samples/nonfree/dmca.pdf'))
+        image_files = self.extract_images(absolute_sample_path('../samples/nonfree/dmca.pdf'))
         assert image_files[0].endswith('bmp')
 
     def test_nonfree_175(self):
         """Extract images of pdf containing jpg images"""
-        self.extract_images(full_path('../samples/nonfree/175.pdf'))
+        self.extract_images(absolute_sample_path('../samples/nonfree/175.pdf'))
 
+    def test_jbig2_image_export(self):
+        """Extract images of pdf containing jbig2 images
 
-if __name__ == '__main__':
-    nose.runmodule()
+        Feature test for: https://github.com/pdfminer/pdfminer.six/pull/46
+        """
+        image_files = self.extract_images(absolute_sample_path('../samples/contrib/pdf-with-jbig2.pdf'))
+        assert image_files[0].endswith('.jb2')
+
+    def test_contrib_matplotlib(self):
+        """Test a pdf with Type3 font"""
+        run('contrib/matplotlib.pdf')
+
+    def test_nonfree_cmp_itext_logo(self):
+        """Test a pdf with Type3 font"""
+        run('nonfree/cmp_itext_logo.pdf')