+++ CHANGELOG.md
@@ -7,6 +7,33 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 Nothing yet
 
+## [20191107] - 2019-11-07
+
+### Deprecated
+- The argument `_py2_no_more_posargs` because Python2 is removed on January
+, 2020 ([#328](https://github.com/pdfminer/pdfminer.six/pull/328) and 
+[#307](https://github.com/pdfminer/pdfminer.six/pull/307))
+
+### Added
+- Simple wrapper to easily extract text from a PDF file [#330](https://github.com/pdfminer/pdfminer.six/pull/330)
+- Support for extracting JBIG2 encoded images ([#311](https://github.com/pdfminer/pdfminer.six/pull/311) and [#46](https://github.com/pdfminer/pdfminer.six/pull/46))
+- Sphinx documentation that is published on 
+  [Read the Docs](https://pdfminersix.readthedocs.io/)
+  ([#329](https://github.com/pdfminer/pdfminer.six/pull/329))
+
+### Fixed
+- Unhandled AssertionError when dumping pdf containing reference to object id 0 
+ ([#318](https://github.com/pdfminer/pdfminer.six/pull/318))
+- Debug flag actually changes logging level to debug for pdf2txt.py and
+ dumppdf.py ([#325](https://github.com/pdfminer/pdfminer.six/pull/325))
+
+### Changed
+- Using argparse instead of getopt for command line interface of dumppdf.py ([#321](https://github.com/pdfminer/pdfminer.six/pull/321))
+- Refactor `LTLayoutContainer.group_textboxes` for a significant speed up in layout analysis ([#315](https://github.com/pdfminer/pdfminer.six/pull/315))
+
+### Removed
+- Files for external applications such as django, cgi and pyinstaller ([#314](https://github.com/pdfminer/pdfminer.six/issues/314))
+
 ## [20191020] - 2019-10-20
 
 ### Deprecated
@@ -27,7 +54,7 @@ Nothing yet
 - Allow for bounding boxes with zero height or width by removing assertion ([#246](https://github.com/pdfminer/pdfminer.six/pull/246))
 
 ### Changed
-- All dependencies are managed in `setup.py` ([#306](https://github.com/pdfminer/pdfminer.six/pull/306), [#219](https://github.com/pdfminer/pdfminer.six/pull/219))
+- All dependencies are managed in `setup.py` ([#306](https://github.com/pdfminer/pdfminer.six/pull/306) and [#219](https://github.com/pdfminer/pdfminer.six/pull/219))
 
 ## [20181108] - 2018-11-08